Refactor signer database using the same procedure as the aggregator
* Providers are split in dedicated file
* Records are moved to a `record` module
* Repositories are moved to a `repository` module
