+++ mithril-signer/src/database/mod.rs
@@ -3,7 +3,9 @@
 //! representation with their associated providers.
 pub mod cardano_transaction_migration;
 pub mod migration;
-pub mod provider;
+pub(crate) mod provider;
+pub mod record;
+pub mod repository;
 
 #[cfg(test)]
 pub mod test_utils {
